requirements_1.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2

# Web Crawling
scrapy==2.11.2
playwright==1.48.0
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3
aiohttp==3.10.5
selenium==4.25.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.4
pdf2image==1.17.0
pytesseract==0.3.13
pypdfium2==4.30.0

# OCR
pytesseract==0.3.13
easyocr==1.7.1

# Vector Database
chromadb==0.5.5
faiss-cpu==1.8.0
sentence-transformers==3.1.1

# LLM APIs
google-generativeai==0.8.3
openai==1.51.2  # For DeepSeek API compatibility

# Machine Learning & Data Science
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
langchain==0.3.3
langchain-community==0.3.2
langchain-google-genai==2.0.0

# Text Processing & NLP
nltk==3.9.1
spacy==3.7.6
ftfy==6.2.3
unidecode==1.3.8

# Multilingual Support
polyglot==16.7.4
transliterate==1.10.2

# Background Tasks & Scheduling
apscheduler==3.10.4
celery==5.4.0
redis==5.1.1

# Database
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0

# Logging & Monitoring
loguru==0.7.2
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
tqdm==4.66.5
validators==0.34.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
